Migration fixes

+++ slipspace.sh
@@ -1,6 +1,6 @@
 function slipspace() {
-  if [ $# -ne 2 ]; then
-    echo "Usage: slipspace deploy [environment]"
+  if [ ! $# -ge 2 ]; then
+    echo "Usage: slipspace deploy [environment] [options]"
   else
     if [ $1 = "deploy" ]; then
       if [ ! -f slipspace.json ]; then
@@ -15,15 +15,15 @@ function slipspace() {
         path="${path//\"}"
         after=`cat slipspace.json | jq -c -r ".environments.$env.after[] + \";\" "`
         migrate=`cat slipspace.json | jq ".environments.$env.migrate // null"`
-        if [ "$migrate" = "null" ]; then
-          ssh -t $user@$host "echo 'Deploying to ${env}...'; cd ${path}; echo 'Fetching source...'; git reset --hard; git checkout ${env}; git pull origin ${env}; echo 'Running post-deploy tasks...'; ${after}exit; bash"
-        else
+        if [ ! "$migrate" = "null" ] && [ ! -z $3 ] && [ $3 = "--migrate" ]; then
           ssh -t $user@$host "echo 'Deploying to ${env}...'; cd ${path}; echo 'Fetching source...'; git reset --hard; git checkout ${env}; git pull origin ${env}; echo 'Running post-deploy tasks...'; ${after}echo 'Running migrations...'; ${migrate}; exit; bash"
+        else
+          ssh -t $user@$host "echo 'Deploying to ${env}...'; cd ${path}; echo 'Fetching source...'; git reset --hard; git checkout ${env}; git pull origin ${env}; echo 'Running post-deploy tasks...'; ${after}exit; bash"
         fi
         echo "Deployment to ${env} has finished."
       fi  
     else
-      echo "Invalid command. Usage: slipspace deploy [environment]"
+      echo "Invalid command. Usage: slipspace deploy [environment] [options]"
     fi  
   fi  
 }
